+++ django_functest/tests/test_common.py
@@ -201,6 +201,11 @@ class TestCommonBase(object):
         self.get_url('django_functest.test_misc')
         self.assertTextPresent("Hello to The Jabberwocky")
 
+    def test_get_session_data(self):
+        self.get_url('django_functest.set_sess_foo_to_bar')
+        sess_dict = self.get_session_data()
+        self.assertEqual(sess_dict, {'foo': 'bar'})
+
     def test_value(self):
         self.get_url('edit_thing', thing_id=self.thing.id)
         self.assertEqual(self.value('#id_name'),